+++ frontend/package.json
@@ -11,6 +11,7 @@
     "react-bootstrap": "^1.0.0",
     "react-dom": "^16.13.0",
     "react-redux": "^7.2.0",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.0",
     "redux": "^4.0.5"
   },
