+++ frontend/src/components/Catalog/Catalog.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default class Catalog extends React.Component{
+  render(){
+    return(
+      <div>
+        <h1>Catalog</h1>
+      </div>
+    )
+  }
+}
