+++ frontend/src/components/MainPage/MainManga.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import '../App.css'
+import './MainStyle.css'
 
 const MangaCard = (props) =>{
   const mangaCards = []
