+++ frontend/src/components/MainPage/MainNews.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import '../App.css'
+import './MainStyle.css'
 
 
 const NewsCard = (props) =>{
