+++ frontend/src/components/MainPage/MainStyle.css
@@ -1,0 +1,292 @@
+
+/* NAVBAR MAIN */
+.Navbar{
+  border-radius: 8px;
+  height: 70px;
+}
+.navbar-brand {
+  color: #fff !important;
+  font-size: 24px !important;
+}
+.nav-link{
+  color: #fff !important;
+  font-size: 20px;
+}
+.NavInput{
+  width: 300px !important;
+  height: 25px !important;
+  border-radius: 10px !important;
+}
+
+
+/* NAVBAR BUTTON */
+.form-switch {
+  display: inline-block;
+  cursor: pointer;
+  -webkit-tap-highlight-color: transparent;
+  margin-top: 10px;
+}
+
+.form-switch i {
+  position: relative;
+  display: inline-block;
+  margin-right: .5rem;
+  width: 46px;
+  height: 26px;
+  background-color: #e6e6e6;
+  border-radius: 23px;
+  vertical-align: text-bottom;
+  transition: all 0.3s linear;
+}
+
+.form-switch i::before {
+  content: "";
+  position: absolute;
+  left: 0;
+  width: 42px;
+  height: 22px;
+  background-color: #11998E;
+  border-radius: 11px;
+  transform: translate3d(2px, 2px, 0) scale3d(1, 1, 1);
+  transition: all 0.25s linear;
+}
+
+.form-switch i::after {
+  content: "";
+  position: absolute;
+  left: 0;
+  width: 22px;
+  height: 22px;
+  background-color: #fff;
+  border-radius: 11px;
+  box-shadow: 0 2px 2px rgba(0, 0, 0, 0.24);
+  transform: translate3d(2px, 2px, 0);
+  transition: all 0.2s ease-in-out;
+}
+
+.form-switch:active i::after {
+  width: 28px;
+  transform: translate3d(2px, 2px, 0);
+}
+
+.form-switch:active input:checked + i::after { transform: translate3d(16px, 2px, 0); }
+
+.form-switch input { display: none; }
+
+.form-switch input:checked + i { background-color: #B52556; }
+
+.form-switch input:checked + i::before { transform: translate3d(18px, 2px, 0) scale3d(0, 0, 0); }
+
+.form-switch input:checked + i::after { transform: translate3d(22px, 2px, 0); }
+
+/* NAVBAR SEARCH */
+
+.flexbox {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.search {
+  margin: 20px;
+}
+
+.search > h3 {
+  font-weight: normal;
+}
+
+.search > h1,
+.search > h3 {
+  color: white;
+  margin-bottom: 15px;
+  text-shadow: 0 1px #0091c2;
+}
+
+.search > div {
+  display: inline-block;
+  position: relative;
+}
+
+.search > div:after {
+  content: "";
+  background: white;
+  width: 4px;
+  height: 15px;
+  position: absolute;
+  top: 20px;
+  right: -2px;
+  transform: rotate(135deg);
+  box-shadow: 1px 0 #0091c2;
+}
+
+.search > div > input {
+  color: white;
+  font-size: 16px;
+  background: transparent;
+  width: 25px;
+  height: 25px;
+  padding: 10px;
+  border: solid 3px white;
+  outline: none;
+  border-radius: 35px;
+  box-shadow: 0 1px #0091c2;
+  transition: width 0.5s;
+}
+
+.search > div > input::placeholder {
+  color: #efefef;
+  opacity: 1;
+}
+
+.search > div > input::-ms-placeholder {
+  color: #efefef;
+}
+
+.search > div > input::-ms-input-placeholder {
+  color: #efefef;
+}
+
+.search > div > input:focus,
+.search > div > input:valid {
+  width: 250px;
+}
+
+
+/* SIDEBAR MAIN */
+
+.sidebar__content{
+  background: rgba(196, 196, 196, 0.3);
+  min-height: 306px;
+  width: 100%;
+  border-radius: 8px;
+  border: 2px solid #fff;
+  text-align: center;
+
+}
+
+.sidebarTitle{
+  font-size: 26px;
+  margin: 20px 0 15px 35px;
+}
+.ads__img{
+  max-height: 260px;
+  display: inline-block;
+  margin-top: 5%;
+}
+
+/* MAIN CONTENT */
+
+.MainContent{
+  background: rgba(196, 196, 196, 0.3);
+  border-radius: 8px;
+  border: 2px solid #fff;
+  margin: 0 20px 0 20px;
+  min-height: 240px;
+}
+.componentTitle{
+  font-size: 26px;
+  margin: 20px 0px 15px 35px; 
+}
+.centered{
+  text-align: center;
+  max-height: 260px;
+  margin-top: 20px;
+}
+.mangaCard{
+  display: inline-block;
+  height: 250px;
+}
+.mangaCard__title{
+  font-size: 15px;
+}
+
+/* MAIN PICTURES */
+
+.mangaCard__img {
+  opacity: 1;
+  display: block;
+  height: 250px;
+  transition: .5s ease;
+  backface-visibility: hidden;
+  border-radius: 8px;
+}
+
+.middle {
+  transition: .5s ease;
+  opacity: 0;
+  position: relative;
+  top: -125px;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  -ms-transform: translate(-50%, -50%);
+  text-align: center;
+}
+
+.mangaCard:hover .image {
+  opacity: 0.3;
+}
+
+.mangaCard:hover .middle {
+  opacity: 1;
+}
+
+.mangaName {
+  background-color: rgba(0, 0, 0, 0.7);
+  color: white;
+  font-size: 20px;
+  width: 179px;
+  height: 250px;
+  padding: 5px 5px 0 5px;
+  border-radius: 8px;
+}
+
+/* MAIN NEWS */
+.MainNews{
+  background: rgba(196, 196, 196, 0.3);
+  border-radius: 8px;
+  border: 2px solid #fff;
+  margin:20px;
+  min-height: 240px;
+  padding: 10px;
+  margin-top: 0;
+}
+
+/* NEWS PICTURES */
+
+.newsCard__img {
+  opacity: 1;
+  display: block;
+  height: 230px;
+  transition: .5s ease;
+  backface-visibility: hidden;
+  border-radius: 8px;
+}
+
+.newsCard:hover .image {
+  opacity: 0.3;
+}
+
+.newsCard:hover .middleNews {
+  opacity: 1;
+}
+
+.middleNews {
+  transition: .5s ease;
+  opacity: 0;
+  position: relative;
+  top: -115px;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  -ms-transform: translate(-50%, -50%);
+  text-align: center;
+}
+
+.newsName {
+  background-color: rgba(0, 0, 0, 0.7);
+  color: white;
+  font-size: 20px;
+  width: 306px;
+  height: 230px;
+  padding: 5px 5px 0 5px;
+  border-radius: 8px;
+}
