+++ frontend/src/components/Rate/Rate.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default class Rate extends React.Component{
+  render(){
+    return(
+      <div>
+        <h1>Rate</h1>
+      </div>
+    )
+  }
+}
